% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc.lpars}
\alias{calc.lpars}
\title{Calculates scaling parameters required by \code{sim.n}}
\usage{
calc.lpars(n, ymin, ymax, W, hfun, b, pi.x, logphi, nx = 100, ny = 100,
  inflate = 1.05)
}
\arguments{
\item{n}{expected sample size reqired from NHPP}

\item{ymin}{smallest forward distance}

\item{ymax}{largest forward distance}

\item{W}{perpendicular truncation distance}

\item{hfun}{detection hazard function}

\item{b}{vector of detection hazard function parameters}

\item{pi.x}{perpendicular distance distribution function}

\item{logphi}{vector with log of pi.x parameters}

\item{nx}{number of intervals on x-axis at which to calculate intensity}

\item{ny}{number of intervals on y-axis at which to calculate intensity}

\item{inflate:}{multiplier by which to increase max intensity in region; this is
       just a safeguard against the max at some point not calculated being
       greater than the points at which it was calculated. (The sample
       generator in \code{rpoispp} uses rejection sampling so needs the
       global max.)}
}
\value{
Object of class "ppscale", being alist with two parameters:
\code{$lscale} is the multiplier needed to make the total intensity equal
to n, while \code{$lmax} is maximum lintensity in the region.
}
\description{
Calculates scaling parameters required by \code{sim.n} in
             order to generate sample of given size.
}
\details{
Calls \code{\link{poisint}} on grid of (x,y) points, calculates total
intensity in area and then calculates scaling needed to make this equal to
n, and the maximum intensity in the region.

Output from this function can be passed as the \code{lscale} argument to
\code{\link{sim.n}}.
}
\examples{
\dontrun{
n=100;ymin=0.01;ymax=5;W=2
b=log(c(0.75,1));logphi=c(0.5,log(0.3))
intscale=calc.lpars(n,ymin,ymax,W,h2,b,pi.norm,logphi)
intscale
}
}

