% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{coveragep}
\alias{coveragep}
\title{Calculate coverage probabilities of \eqn{\hat p} for simulated data}
\usage{
coveragep(fit, true.hr, true.b, true.pi.x, true.logphi, type = "LOGNORM",
  interval = 0.95, verbose = FALSE)
}
\arguments{
\item{fit}{object resulting from a call of \code{\link{fityx}} (see details)}

\item{true.hr}{true form of hazard rate function}

\item{true.b}{true values of hazard rate parameters}

\item{true.pi.x}{true form of perpendicular density distribution pi(x)}

\item{true.logphi}{true values for pi(x) perpendicular density distribution parameters}

\item{type}{\code{LOGNORM} log-normal confidence intervals; \code{NORM} normal confidence intervals.}

\item{interval}{the interval used to determine coverage probability}

\item{verbose}{boolean.  FALSE only covered returned.  TRUE a data frame of covered and associated calculations returned (see returns).}
}
\value{
boolean; \code{TRUE} if within log-normal confidence interval, otherwise \code{FALSE}.
\code{verbose=TRUE}   a data frame of p phat, and CV[phat], interval, lower bound returned.
}
\description{
Calculate coverage probabilities of \eqn{\hat p} using the delta method and assuming a 
log-normal error distribution.
}
\details{
In the call of \code{\link{fityx}} must have \code{hessian=TRUE}
}
\examples{
\dontrun{
ystart=4;w=1
hr=h2; b=log(c(0.75,1))
pi.x=pi.norm; logphi=c(0.5,log(0.2))
N=50 #true number of animals
#generate some observations
simDat=simXY(N=N,pi.x=pi.x,logphi=logphi,
hr=hr,b=b,w=w,ystart=ystart)
x=simDat$locs$x; y=simDat$locs$y 
est.yx=fityx(y,x,b,hr,ystart,pi.x,logphi,w)
coveragep(fit=est.yx,true.hr=hr,true.b=b,interval=0.95,
 true.pi.x=pi.x,true.logphi=logphi,verbose=TRUE)
}     
}
\seealso{
\code{\link{phat}} \code{\link{fityx}}
}

