% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{dolphin.dat}
\alias{dolphin.dat}
\title{Dolphin line transect survey data.}
\format{A list with elements x (perpendicular distance) and y (forward distance).}
\source{
We are gerateful to North Atlantic Marine Mammal Commission (NAMMCO) and the 
Faroese Museum of Natural History for allowing us to use the dolphin survey data from 
1995 North Atlantic Sightings Survey (NASS95). These data are analysed using 
mark-recapture distance sampling methods by Ca\~{n}adas et al. (2004).
}
\usage{
dolphin.dat
}
\description{
Locations relative to the observer of 74 detections of dolphins from a 
shipboard visual survey.
}
\examples{
 data(primate.dat)
}
\references{
Ca\~{n}adas, A., Desportes, G. and Borchers, D.L. 2004. The estimation of the detection 
function and g(0) for short-beaked common dolphins (Delphinis delphis), using 
double-platform data collected during the NASS-95 Faroese survey. Journal of Cetacean 
Research and Management 6: 191-198.
}

