% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{ep1}
\alias{ep1}
\title{Three-parameter exponential power hazard detection function}
\usage{
ep1(y, x, b)
}
\arguments{
\item{y}{Forward distance}

\item{x}{perpendicular distance}

\item{b}{parameter vector, where \code{b[1]} is plogis(theta[1])  \code{b[2]} is 
log(theta[2]) and \code{b[3]} is log(b[3]).}
}
\value{
probability of detection given that an animal is availabe at location x,y
#'@examples
b=c(1, -4, 1)
ep1(0.5,0.5,b=b)
yy=seq(0,0.03,length=100);xx=rep(0,100)
hh=ep1(yy,xx,b=b)
plot(yy,hh,type="l")
}
\description{
Inverse power hazard function, as per Borchers and Langrock (in press):
Has form h(y,x)=theta[1]*exp(-(x^theta[3]+y^theta[3])/(theta[2]^theta[3])).
}
\references{
Borchers, D.L and Langrock, R."Double-observer line transect surveys with Markov-
modulated Poisson process models for animal availability" Biometrics (in press).
}

