% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fityx}
\alias{fityx}
\title{Maximum likelihood estimation for unknown hazard and perpendicular distance distribution}
\usage{
fityx(y, x, b, hr, ystart, pi.x, logphi, w, control = list(),
  hessian = FALSE, corrFlag = 0.7, ...)
}
\arguments{
\item{y}{scalar or vector; forward distance observations}

\item{x}{scale or vector; perp. distance observations}

\item{b}{two-element vector of hazard rate parameters, some of whihc may be logged}

\item{hr}{hazard rate function}

\item{ystart}{max forward distance at which could possibly detect animal (see details).}

\item{pi.x}{perpendicular distance density distribution}

\item{logphi}{parameters for pi.x (some maybe logged)}

\item{w}{perpendicular truncation distance.}

\item{control}{see \code{\link{optim}} control}

\item{hessian}{return hessian.  See also \code{\link{optim}}.}

\item{...}{arguments to be passed into \code{\link{optim}}}

\item{corrFlag=0.7}{Absolute parameter correlation value above which a warning is issued.}
}
\value{
\code{\link{optim}} fit object and
\code{$hr} = hazard rate function used.
\code{$pi.x} = perpendicular distance function used.
\code{$ystart} = ystart max forward distance detection used.
\code{$w} = perpendicular truncation distance used.
\code{$b} = estimated hazard parameters (\code{$par[1:2]})
\code{$logphi} (\code{$par[3:4]})
\code{AIC} AIC value
And if \code{hessian=TRUE}
\code{vcov} variance covariance matrix.  Will warn if there is a problem inverting the hessian.
\code{CVpar} Coefficient of variation for each paramter estimate.
\code{error} Boolean, \code{TRUE} if convergence!=0 or problem inverting the hessian, or parameter correlation is exceeded.
}
\description{
Uses \code{\link{optim}} to obtain a MLE for the hazard function and animal
perpendicular distance distribution.  Functional forms for the hazard and perpendicular distance
distribution must be specified.
}
\details{
Must to ensure the hazard function has decayed to (very close to) zero by \code{ystart}.
}
\examples{
\dontrun{
ystart=4;w=1
hr=h2; b=log(c(0.75,1))
pi.x=pi.norm; logphi=c(0.5,log(0.2))
N=50 #true number of animals
#generate some observations
simDat=simXY(N=N,pi.x=pi.x,logphi=logphi,
hr=hr,b=b,w=w,ystart=ystart)
x=simDat$locs$x; y=simDat$locs$y
fityx(fityx(y,x,b,hr,ystart,pi.x,logphi,w)
}
}
\seealso{
\code{\link{negloglik.yx}}
}

