% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLToptimx.R
\name{fityxOptimx}
\alias{fityxOptimx}
\title{Maximum likelihood estimation for unknown hazard and perpendicular distance distribution}
\usage{
fityxOptimx(y, x, b, hr, ystart, pi.x, logphi, w, method = "Nelder-Mead",
  lower = -Inf, upper = Inf, control = list(), itnmax = NULL,
  hessian = FALSE, corrFlag = 0.7, ...)
}
\arguments{
\item{y}{forward distance observations}

\item{x}{perpendicular distance observations}

\item{b}{two to four element vector of hazard rate parameters, some of which may be logged}

\item{hr}{hazard rate function}

\item{ystart}{max forward distance at which could possibly detect animal (see details).}

\item{pi.x}{perpendicular distance density distribution}

\item{logphi}{parameters for pi.x (some maybe logged)}

\item{w}{perpendicular truncation distance.}

\item{method}{optimisation method to be used by \code{\link{optimx}}. Default is \code{"Nelder-Mead"}}

\item{lower, }{upper Bounds for parameters for use with methods such as \code{"L-BFGS-B"}. See \code{\link{optimx}}.}

\item{control}{see \code{\link{optimx}} control}

\item{itnmax}{maximum number of iterations for the \code{Nelder-Mead} method.  NB this is not passed in via the control argument}

\item{hessian}{return hessian.  See also \code{\link{optimx}}.}

\item{...}{arguments to be passed into \code{\link{optimx}}}

\item{corrFlag=0.7}{Absolute parameter correlation value above which a warning is issued.}
}
\value{
\code{\link{optim}} fit object and \cr
\code{$hr} = hazard rate function used.\cr
\code{$pi.x} = perpendicular distance function used.\cr
\code{$ystart} = ystart max forward distance detection used.\cr
\code{$w} = perpendicular truncation distance used.\cr
\code{$b} = estimated hazard parameters\cr
\code{$dat} = data frame with data (\code{$x} and \code{$y})\cr
\code{$logphi} \cr
\code{AIC} AIC value\cr
And if \code{hessian=TRUE}:\cr
\code{vcov} variance covariance matrix.  Will warn if there is a problem inverting 
the hessian.\cr
\code{CVpar} Coefficient of variation for each paramter estimate. \cr
\code{error} Boolean, \code{TRUE} if convergence!=0 or problem inverting the hessian, 
or parameter correlation is exceeded.\cr
}
\description{
Uses \code{\link{optim}} to obtain a MLE for the hazard function and animal
perpendicular distance distribution.  Functional forms for the hazard and perpendicular distance 
distribution must be specified.
}
\details{
Must to ensure the hazard function has decayed to (very close to) zero by 
\code{ystart}.
}
\examples{
\dontrun{
#Data preparation:
w=0.03;ystart=0.05
library(xlsx)
dat=read.xlsx("~/Dropbox/packages/LT2D/data/Jantho Primate Line.xlsx",1)
x=dat$PP.Distance
y=dat$Forward.Distance
nas=which(is.na(y))
x=x[-nas]
y=y[-nas]
gtw=which(x>w)
x=x[-gtw]
y=y[-gtw]
#Example model fits:

b=c(-7.3287948, 0.9945317)
logphi=c(.01646734, -4.67131261)
fit.n.optx=NULL
fit.n.optx=fityxOptimx(y,x,b=b,
           hr=h1,ystart=ystart,
           control=list(trace=5),hessian=TRUE,
           pi.x=pi.norm,logphi=logphi,w=w)

b=c(-7.3329141,0.9948721)
logphi=c(-0.05,-4.7)
fit.chn.optx=NULL
fit.chn.optx=fityxOptimx(y,x,b=b,hr=h1,ystart=ystart,
             pi.x=pi.chnorm,logphi=logphi,w=w,itnmax=5000,
             hessian=TRUE,control=list(trace=5))

b=c(5.2919208, -0.2205593, 8.4701307)
logphi=c(0.01784102, -4.42209067)
fit.n.ip1.optx=NULL
fit.n.ip1.optx=fityxOptimx(y,x,b=b,hr=ip1,ystart=ystart,
               pi.x=pi.norm,logphi=logphi,w=w,
               hessian=TRUE,control=list(trace=6))
}
}
\seealso{
\code{\link{negloglik.yx}} \code{\link{fityx}} \code{\link{optimx}}
}

