% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{fyx}
\alias{fyx}
\title{Waiting distance pdf}
\usage{
fyx(y, x, b, hr, ystart, nint = 100)
}
\arguments{
\item{y}{scalar or vector; forward distance}

\item{x}{scale or vector; perp. distance}

\item{b}{two-element vector of hazard rate parameters, some of whihc may be logged}

\item{hr}{hazard rate function}

\item{ystart}{max forward distance at which could possibly detect animal (see details).}

\item{nint}{number of intervals in numerical integration.}
}
\value{
pdf of waiting distance at x,y
}
\description{
Calculates the pdf of the 'waiting distance' \eqn{f(y,x)=h(y,x)*\exp(-\int_y^{ystart} h(t,x) dt)}.
}
\details{
Need to ensure the hazard function has decayed to (very close to) zero by \code{ystart}.
}
\examples{
w=1; ystart=4
gridx=seq(0,w,length=50); gridy=seq(0,ystart,length=50)
f=outer(gridy,gridx,FUN=fyx,b=c(-0.2876821, -2.3025851),hr=h2,ystart)
persp(gridx,gridy,t(f),theta=45,phi=35,zlab="f(y|x)")
}

