% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{ghy}
\alias{ghy}
\title{Detection hazard function \code{ghy} prob(detect | available at x,y)}
\usage{
ghy(y, x, b)
}
\arguments{
\item{y}{Forward distance}

\item{x}{perpendicular distance}

\item{b}{parameter vector, where \code{b[2]} is log(theta); theta[1] is as per \link{\code{h1}}
parameter theta[1]; theta[2] is as per \link{\code{h1}} theta[2]; theta[3] is parameter that
is added to y to shift forward distance origin and allow p(0)<1.}
}
\value{
probability of detection given that an animal is availabe at location x,y
}
\description{
This hazard function is a generalization of the form k(r,y)=a*r^(-b) from
Hayes and Buckland (1983) p36, the generalization being that a parameter to be estimated
is added to y. When this parameter is zero you get the form of Hayes and Buckland (1983) p36.
Note: This function uses x for perp. dist., they use y.
}
\examples{
h1(0.5,0.5,b=log(c(0.001,1)))
ghy(0.5,0.5,b=log(c(0.001,1,0.0)))
ghy(0.5,0.5,b=log(c(0.001,1,0.01)))
}
\references{
Hayes, R. J., and S. T. Buckland. "Radial-distance models for the line-transect method." Biometrics (1983): 29-42.
}
\seealso{
\code{\link{h2}}, \code{\link{ghy2}}
}

