% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{ghy2}
\alias{ghy2}
\title{Detection hazard function \code{ghy2} prob(detect | available at x,y)}
\usage{
ghy2(y, x, b)
}
\arguments{
\item{y}{Forward distance}

\item{x}{perpendicular distance}

\item{b}{parameter vector, where \code{b[2]} is log(theta); theta[1] is as per \link{\code{h1}}
parameter theta[1]; theta[2] is as per \link{\code{h1}} theta[2]; theta[3] is parameter that
is added to y to shift forward distance origin and allow p(0)<1; theta[4] is the equivalent of
theta[1], but specific to y, whereas theta[1] is specific to x in this function.}
}
\value{
probability of detection given that an animal is availabe at location x,y
}
\description{
This hazard function is a generalization of the form k(r,y)=a*r^(-b) from
Hayes and Buckland (1983) p36, the generalization being that (1) a parameter to be estimated
is added to y, and (2) x and y have separate scale parameters. It is a generalization of
\link{\code{ghy}} to allow x and y to have separate scale parameters.
Note: This function uses x for perp. dist., they use y.
}
\examples{
h1(0.5,0.5,b=log(c(0.001,1)))
ghy(0.5,0.5,b=log(c(0.001,1,0.0)))
ghy(0.5,0.5,b=log(c(0.001,1,0.01)))
ghy2(0.5,0.5,b=log(c(0.001,1,0.01,0.001)))
ghy2(0.5,0.5,b=log(c(0.001,1,0.01,0.005)))
}
\references{
Hayes, R. J., and S. T. Buckland. "Radial-distance models for the line-transect method." Biometrics (1983): 29-42.
}
\seealso{
\code{\link{h2}}, \code{\link{ghy2}}
}

