% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTGoF.R, R/GoF.r
\name{GoFx}
\alias{GoFx}
\title{Goodness-of-fit in perpendicular dimension.}
\usage{
GoFx(fit, plot = FALSE, nint = 100, dotitle = FALSE)

GoFx(fit, plot = FALSE, nint = 100, dotitle = FALSE)
}
\arguments{
\item{plot}{If TRUE, does Q-Q plot. Point corresponding to largest difference between
empirical and theoretical cdf (on which the Kolmogarov-Smirnov test is based) is circled in red.}

\item{hmltm}{fitted model, as output by \code{\link{est.hmltm}}}

\item{hmltm}{fitted model, as output by \code{\link{est.hmltm}}}

\item{plot}{If TRUE, does Q-Q plot. Point corresponding to largest difference between
empirical and theoretical cdf (on which the Kolmogarov-Smirnov test is based) is circled in red.}
}
\value{
data frame with these elements
\code{$p.cvm} = Cramer-von Mises p-value.
\code{$D.kolomogarov} = x value of Kolmogarov-distributed random variable
\code{$p.kolomogarov} = kolomogarov p-value (which is
based on only the largest difference between emprical and theoretical cdfs).
\code{$qq.x} = empirical distribution function values.
\code{$qq.y} = cumulative distribution function values.
\code{$x} = x values.

data frame with these elements
\code{$p.cvm} = Cramer-von Mises p-value.
\code{$D.kolomogarov} = x value of Kolmogarov-distributed random variable
\code{$p.kolomogarov} = kolomogarov p-value (which is
based on only the largest difference between emprical and theoretical cdfs).
\code{$qq.x} = empirical distribution function values.
\code{$qq.y} = cumulative distribution function values.
\code{$x} = x values.
}
\description{
Calculates goodness-of-fit in perpendicular dimension, plots fit, and returns p-value and 
other stuff. Returns two p-values: \code{p.ks} is the Kolmogarov-Smirnov p-value (which is
based on only the largest difference between emprical and theoretical cdfs), and Cramer-von Mises
p-value (which is based on all cdf values).

Calculates goodness-of-fit in perpendicular dimension, plots fit, and returns p-value and
other stuff. Returns two p-values: \code{p.ks} is the Kolmogarov-Smirnov p-value (which is
based on only the largest difference between emprical and theoretical cdfs), and 
Cramer-von Mises p-value (which is based on all cdf values).
}
\examples{
\dontrun{
ystart=0.05;w=0.03
logphi=c(0.0180552, -4.4215995)
b=c(-23.725809,  -3.136638  , 2.122910)
N=200 #true number of animals
#generate some observations
simDat=simXY(N=N,pi.x=pi.norm,logphi=logphi,hr=ip1,b=b,w=w,ystart=ystart)

x=simDat$locs$x; y=simDat$locs$y 
est.yx=fityx(y,x,b,hr=ip1,ystart,pi.x=pi.norm,logphi,w) 
plotfit.x(x=x,est=est.yx)
rug(x=est.yx$dat$x)
tst=GoFx(fit=est.yx,plot=TRUE)
}
\dontrun{
ystart=0.05;w=0.03
logphi=c(0.0180552, -4.4215995)
b=c(-23.725809,  -3.136638  , 2.122910)
N=200 #true number of animals
#generate some observations
simDat=simXY(N=N,pi.x=pi.norm,logphi=logphi,
hr=ip1,b=b,w=w,ystart=ystart)

x=simDat$locs$x; y=simDat$locs$y
est.yx=fityx(y,x,b,hr=ip1,ystart,pi.x=pi.norm,logphi,w)
plotfit.x(x=x,est=est.yx)
rug(x=est.yx$dat$x)
tst=GoFx(fit=est.yx,plot=TRUE)
}
}
\seealso{
\code{\link{fityx}} \code{\link{p.kolomogarov}} \code{\link{GoFy}}

\code{\link{fityx}} \code{\link{p.kolomogarov}} \code{\link{GoFy}}
}

