% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2DLTGoF.R, R/GoF.r
\name{GoFy}
\alias{GoFy}
\title{Goodness-of-Fit in the forward direction (y)}
\usage{
GoFy(fit, plot = FALSE, dotitle = FALSE)

GoFy(fit, plot = FALSE, dotitle = FALSE)
}
\arguments{
\item{fit}{object resulting from a call of \link{fityx}.}

\item{plot=FALSE}{boolean TRUE - display Q-Q plot.}

\item{fit}{object resulting from a call of \link{fityx}.}

\item{plot=FALSE}{boolean TRUE - display Q-Q plot.}
}
\value{
data frame with these elements
\code{$p.cvm} = Cramer-von Mises p-value.
\code{$D.kolomogarov} = x value of Kolmogarov-distributed random variable
\code{$p.kolomogarov} = kolomogarov p-value.

data frame with these elements
\code{$p.cvm} = Cramer-von Mises p-value.
\code{$D.kolomogarov} = x value of Kolmogarov-distributed random variable
\code{$p.kolomogarov} = kolomogarov p-value.
}
\description{
Calculates the Goodness-of-Fit in the forward direction (y)

Calculates the Goodness-of-Fit in the forward direction (y)
}
\details{
Calculates goodness-of-fit for forward distances and calculates p-values using kolomogarov and Cramer-von Mises.

Calculates goodness-of-fit for forward distances and calculates p-values
using kolomogarov and Cramer-von Mises.
}
\examples{
\dontrun{
ystart=0.05;w=0.03
logphi=c(0.0180552, -4.4215995)
b=c(-23.725809,  -3.136638  , 2.122910)
N=200 #true number of animals
#generate some observations
simDat=simXY(N=N,pi.x=pi.norm,logphi=logphi,
hr=ip1,b=b,w=w,ystart=ystart)

x=simDat$locs$x; y=simDat$locs$y
est.yx=fityx(y,x,b,hr=ip1,ystart,pi.x=pi.norm,logphi,w)
plotfit.y(y,x,est.yx,nclass=10)
GoFy(fit=est.yx,plot=TRUE)
}
\dontrun{
ystart=0.05;w=0.03
logphi=c(0.0180552, -4.4215995)
b=c(-23.725809,  -3.136638  , 2.122910)
N=200 #true number of animals
#generate some observations
simDat=simXY(N=N,pi.x=pi.norm,logphi=logphi,
hr=ip1,b=b,w=w,ystart=ystart)

x=simDat$locs$x; y=simDat$locs$y
est.yx=fityx(y,x,b,hr=ip1,ystart,pi.x=pi.norm,logphi,w)
plotfit.y(y,x,est.yx,nclass=10)
GoFy(fit=est.yx,plot=TRUE)
}
}
\seealso{
\code{\link{fityx}}, \code{\link{}}

\code{\link{fityx}} \code{\link{p.kolomogarov}} \code{\link{GoFx}}

\code{\link{fityx}}, \code{\link{}}

\code{\link{fityx}} \code{\link{p.kolomogarov}}
\code{\link{GoFx}}
}

