% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h.exp2}
\alias{h.exp2}
\title{Detection hazard function \code{h.exp2} prob(detect | available at x,y)}
\usage{
h.exp2(y, x, b = c(0, 0))
}
\arguments{
\item{y}{Forward distance}

\item{x}{perpendicular distance}

\item{b}{parameter vector, may be logged parameter values}
}
\value{
probability of detection given that an animal is availabe at location x,y
}
\description{
2-paramter Exponential power hazard model of Skaug & Schweder 1999. The gamma parameter is fixed at 2.
}
\examples{
h.exp2(0.5,0.5,b=log(c(0.75,0.9)))
}
\references{
Skaug, Hans J., and Tore Schweder. "Hazard models for line transect surveys with independent observers." Biometrics 55.1 (1999): 29-36.
}
\seealso{
\code{\link{h1}} \code{\link{h2}}
}

