% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{h.okamura}
\alias{h.okamura}
\title{Detection hazard function of Hiroshi et al. (2003)
#'}
\usage{
h.okamura(y, x, b = c(0, 0))
}
\arguments{
\item{y}{Forward distance}

\item{x}{perpendicular distance}

\item{b}{parameter vector, may be logged parameter values}
}
\value{
probability of detection given that an animal is available at location x,y
}
\description{
2-parameter hazard model of Hiroshi, et al. 2003.
}
\examples{
h.okamura(0.5,0.5,b=log(c(0.5,0.9)))
}
\references{
Okamura, Hiroshi et al. (2003) Abundance Estimation of Diving Animals by the Double-Platoform Line Transect Method, Biometrics 59(3):512-520.
}
\seealso{
\code{\link{h1}} \code{\link{h2}} \code{\link{h.exp2}}
}

