% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h1}
\alias{h1}
\title{Detection hazard function \code{h1} prob(detect | available at x,y)}
\usage{
h1(y, x, b)
}
\arguments{
\item{y}{Forward distance}

\item{x}{perpendicular distance}

\item{b}{parameter vector, where \code{b[2]} is log(theta)}
}
\value{
probability of detection given that an animal is availabe at location x,y
}
\description{
This hazard function has the form k(r,y)=a*r^(-b) from Hayes and Buckland (1983) p36.
}
\examples{
h1(0.5,0.5,b=log(c(0.001,1)))
}
\references{
Hayes, R. J., and S. T. Buckland. "Radial-distance models for the line-transect method." Biometrics (1983): 29-42.
}
\seealso{
\code{\link{h2}}
}

