% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{h21}
\alias{h21}
\title{Hazard detection function of form \code{h2} with g(0)<1}
\usage{
h21(y, x, b)
}
\arguments{
\item{y}{Forward distance}

\item{x}{perpendicular distance}

\item{b}{parameter vector, where \code{b[2]} is log(theta) and \code{b[3]} is logit(b[3]) detectability at the observer i.e. \eqn{g(0)=c}}
}
\value{
probability of detection given that an animal is availabe at location x,y
#'
}
\description{
This hazard function has the form \eqn{k(r,y)=c*[a*\sqrt(r^2-y^2)/r^{(b+1)}]; b>2} modified
from Hayes and Buckland (1983) p37, where c is imperfect detectability i.e. \eqn{g(0)<1}.
}
\examples{
h21(0.5,0.5,b=c(log(c(0.75,1)),qlogis(0.9)))
}
\references{
Hayes, R. J., and S. T. Buckland. "Radial-distance models for the line-transect method." Biometrics (1983): 29-42.
}

