% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoF.r
\name{HBhr}
\alias{HBhr}
\title{Hayes+Buckland hazard rate model.}
\usage{
HBhr(x, theta)
}
\arguments{
\item{x}{perpendicular distance(s).}

\item{theta}{parameter vector with theta[1] being HB model parameter a1 and theta[2] being b.}
}
\value{
g(x)=1-exp(-a1*x^(-(b-1))).
}
\description{
Evaluates the Hayes and Buckland (1983) hazard rate model (HB model) as parameterised on 
page 36 of that paper.
}
\examples{
xx=seq(0,0.03,length=100)
p=HBhr(xx,h1.to.HB(c(-7, 0.85)))
plot(xx,p,type="l",xlab="Perpendicular distance (x)",ylab="p(x)",ylim=c(0,1))
}
\seealso{
\code{\link{h1.to.HB}}
}

