% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{hr.to.p}
\alias{hr.to.p}
\title{Calculate HR perp dist function from hazard}
\usage{
hr.to.p(x, b, hr)
}
\arguments{
\item{x}{perpendicular distance.}

\item{b}{vector of parameters of \code{\link{h2}} (on log scale).}

\item{hr}{hazard rate function name (must be character): only "h1" and "h2" valid}
}
\value{
value of hazard rate detection function at x
}
\description{
Implements the g(y) on page 36 or that on page 37 of Hayes and Buckland (1983) 
from the hazard function given on page 37 or 38, respectively, of that paper (and implemented 
in function \code{\link{h2}}).
}
\examples{
b=log(c(0.75,0.1))
x=seq(0,1,length=50)
plot(x,hr.to.p(x,b,hr='h2'),type="l",ylim=c(0,1),xlab="Perpendicular distance",ylab="P(detect)")
}

