% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hr1.to.p}
\alias{hr1.to.p}
\title{Calculate harzard rate perpendicular distance function from hazard rate \code{h1}}
\usage{
hr1.to.p(x, b, w)
}
\arguments{
\item{x}{perpendicular distance.}

\item{b}{vector of parameters of \code{\link{h1}} (on log scale).}

\item{w}{perpendicular truncation distance}
}
\value{
value of hazard rate detection function at x
}
\description{
Implements the a prependicular distance function of the hazard rate function
k(x,y)=a*r^(-b) on page 36  of Hayes and Buckland (1983)
from the hazard function given on page 37 of that paper (and implemented
in function \code{\link{h1}}).
}
\examples{
b=log(c(0.01,1.01))
x=seq(0,1,length=50)
plot(x,hr1.to.p(x,b),type="l",ylim=c(0,1),xlab="Perpendicular distance",ylab="P(detect)")
}
\seealso{
\code{\link{h1}} \code{\link{pi.hr1}}
}

