% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{hr2.to.p}
\alias{hr2.to.p}
\title{Calculate harzard rate perpendicular distance function from hazard rate \code{h2}}
\usage{
hr2.to.p(x, b, w)
}
\arguments{
\item{x}{perpendicular distance.}

\item{b}{vector of parameters of \code{\link{h2}} (on log scale).}

\item{w}{perpendicular truncation distance}
}
\value{
value of hazard rate detection function at x
}
\description{
Implements the a prependicular distance function of the hazard rate function
\eqn{k(r,y)=a \sqrt(r^2-y^2)/r^{(b+1)}; b>2} on page 37  of Hayes and Buckland (1983) 
from the hazard function given on page 38 of that paper (and implemented 
in function \code{\link{h2}}).
}
\examples{
x=seq(0,1,length=50)
plot(x,hr2.to.p(x,b=c(-0.2876821, -2.3025851),w=1),ylim=c(0,1),
xlab='Perp. distance, x', ylab='P(detect)',type='l')
}

