% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{ip1}
\alias{ip1}
\title{Inverse power hazard detection function (with g(0)<1)}
\usage{
ip1(y, x, b)
}
\arguments{
\item{y}{Forward distance}

\item{x}{perpendicular distance}

\item{b}{parameter vector, where \code{b[1]} is inv.logit(theta[1])  \code{b[2]} is
log(theta[2]) and \code{b[3]} is log(b[3]).}
}
\value{
probability of detection given that an animal is availabe at location x,y
#'
}
\description{
Inverse power hazard function, as per Borchers and Langrock (in press):
Has form h(y,x)=theta[1]*[theta[2]/(sqrt{theta[2]^2+x^2+y^2})]^(theta[3]+1).
}
\examples{
ip1(0.5,0.5,b=c(log(c(0.75,1)),qlogis(0.9)))
yy=seq(0,0.03,length=100);xx=rep(0,100)
hh=ip1(yy,xx,b=c(logit(0.5),log(c(0.025,4))))
plot(yy,hh,type="l")
}
\references{
Borchers, D.L and Langrock, R."Double-observer line transect surveys with Markov-
modulated Poisson process models for animal availability" Biometrics (in press).
}

