% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LT2D-package.r
\docType{package}
\name{LT2D-package}
\alias{LT2D}
\alias{LT2D-package}
\title{Line transect estimation in two dimensions}
\description{
Density and abundance estimation using line transect survey data with forward 
distances to detections (assuming continuous availability).
}
\details{
\tabular{ll}{ Package: \tab LT2D \cr Type: \tab Package \cr 
  Version: \tab 1.0 \cr Date: \tab 2015-01-11 \cr License: \tab GNU General 
  Public License Version 2 or later \cr }
}
\author{
David Borchers \email{dlb@st-andrews.ac.uk} and 
Martin Cox \email{martinjamescox@gmail.com}
  
  Maintainer: David Borchers \email{dlb@st-andrews.ac.uk}
}

