% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{LT2D.fit}
\alias{LT2D.fit}
\title{Top level fitting and abundance estimation function for LT2D user}
\usage{
LT2D.fit(DataFrameInput, hr, b, ystart, pi.x, logphi, w, formulas = NULL,
  ipars = NULL, xpars = NULL, ypars = NULL, rmin = 0,
  control = list(), hessian = FALSE, corrFlag = 0.7, debug = FALSE)
}
\arguments{
\item{DataFrameInput, }{data.frame ; data frame with required columns
 stratum (stratum number of observation), transect (transect number of
observation, object(object number of detection, NA if transect with no
detections), size (group size of detection), area (stratum area of observation)
and L (transect length of transect).}

\item{hr}{character; name of hazard rate to fit}

\item{b}{numeric; vector of start parameters for hr}

\item{ystart}{numeric; furthest possible forwards distance at which we can
detect animals.}

\item{pi.x}{character; name of perpendicular density to fit}

\item{logphi}{numeric; vector of start parameters for pi.x}

\item{w}{numeric; perpendicular truncation distance}

\item{rmin}{numeric; radial distance below which all values were rounded down to 0}

\item{hessian}{boolean}

\item{corrFlag}{numeric; value above which correlation flag is raised}
}
\description{
This function is a wrapper for the fityx function. It takes
in a 'Distance' style data frame, type-checks it, finds the MLEs using
\code{\link{fityx}} and then calls an abundance function \code{\link{NDest}}
to obtain density and abundance estimates by stratum.
}
