% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{modSelect}
\alias{modSelect}
\title{AIC-based model selection for models fitted using \link{fitxy}}
\usage{
modSelect(modList, modNames = NULL, tab = FALSE, digits = 2,
  maxblength = NULL, maxlogphilength = NULL)
}
\arguments{
\item{modList}{a list of \link{fitxy}-type models.}

\item{modNames}{a vector of model names.}

\item{tab}{boolean - return a data frame suitable for generating a table for report or paper.}

\item{digits}{- number of digits of rounding (see \link{round}) in the table}

\item{maxblength=NULL;}{maximum number of parameters for the hazard function.  If NULL 
the maximum number of parameters for the models in modList is used.}

\item{maxlogphilength=NULL;}{maximum number of parameters for the perpendicular density function.  If NULL 
the maximum number of parameters for the models in modList is used.}
}
\value{
tab=FALSE list; element 1 =data frame of AIC-ranked models, with parameter estimates, 
coefficients of variation and AIC; element 2= vector of the order of AIC-ranked models. tab=TRUE, a three element list with AIC-ranked model order and two tables, the first as tab=FALSE, the 
second, a table for use in reports or manuscripts.
}
\description{
AIC-based model selection of a list of models fitted using \link{fitxy}.  Parameter 
estimates for the hazard and perpendicular distribution functions along with 
associated coefficients of variation are also provided. 
Models are ranked in order of increasing dAIC.  The model list can also be 
returned in a data frame suitable making a tex-type table.
}
\examples{
\dontrun{
ystart=4;w=1
hr=h2; b=log(c(0.75,1))
pi.x=pi.norm; logphi=c(0.5,log(0.2))
N=50 #true number of animals
#generate some observations
simDat=simXY(N=N,pi.x=pi.x,logphi=logphi,
hr=hr,b=b,w=w,ystart=ystart)
x=simDat$locs$x; y=simDat$locs$y 
fit1=fityx(y,x,b,hr,ystart,pi.x,logphi,w)
fit2=fityx(y,x,b=log(c(0.001,1)),hr=h1,ystart,pi.x,logphi,w)
modSelect(modList=list(fit1,fit2),modNames=c('h1','h2'),tab=TRUE)
}
}
\seealso{
\link{fitxy}
}

