% Generated by roxygen2 (4.0.2): do not edit by hand
\name{negloglik.yx}
\alias{negloglik.yx}
\title{Negative log-likelihood for forward distance and perpendicular distance}
\usage{
negloglik.yx(y, x, pars, hr, ystart, pi.x, w, length.b = 2)
}
\arguments{
\item{y}{scalar or vector; forward distance observations}

\item{x}{scale or vector; perp. distance observations}

\item{pars}{c(b,logphi); hazard rate and density log-parameters in a vector (see details).}

\item{hr}{hazard rate function}

\item{ystart}{max forward distance at which could possibly detect animal (see details).}

\item{pi.x}{perpendicular distance density distribution}

\item{w}{perpendicular truncation distance.}

\item{length.b}{length of the hazard rate parameter vector}
}
\value{
negative log likelihood for forward distance, \code{y} and perpendicular distance \code{x}.
}
\description{
Calculates the negative log-likelihood for forward distance, \code{y}, and
perpendicular distance, \code{x}, for a given hazard and perpendicular density distribution.
}
\details{
Must to ensure the hazard function has decayed to (very close to) zero by \code{ystart}.
The parameter vector, \code{pars}, must be passed in with two parameters for the hazard rate first,
then two parameters for perpendicular density gradient \eqn{\pi(x)} i.e. \code{c(b,logphi)}.
}
\examples{
ystart=4;w=1
hr=h2; b=log(c(0.75,1))
pi.x=pi.norm; logphi=c(0.5,log(0.2))
N=50 #true number of animals
#generate some observations
simDat=simXY(N=N,pi.x=pi.x,logphi=logphi,
hr=hr,b=b,w=w,ystart=ystart)
x=simDat$locs$x; y=simDat$locs$y
pars=c(b,logphi)
negloglik.yx(y,x,pars,hr,ystart,pi.x,w)
}
\seealso{
\code{\link{simXY}}
}

