% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openGraph.R
\name{openGraph}
\alias{openGraph}
\title{Function to open an operating system specific graphics device.}
\usage{
openGraph(width = 7, height = 7, ...)
}
\arguments{
\item{width, }{height the (nominal) width and height of the canvas of the plotting window in inches}

\item{...}{other arguments passed into graphics device}
}
\description{
Function to open an operating system specific graphics device.
}
\note{
function downloaded from \link{http://doingbayesiandataanalysis.blogspot.com.au/2013/01/uniform-r-code-for-opening-saving.html}
}
\author{
John K. Kruschke
}

