% Generated by roxygen2 (4.0.2): do not edit by hand
\name{p.pi.x}
\alias{p.pi.x}
\title{Product of p(x) and pi(x)}
\usage{
p.pi.x(x, b, hr, ystart, pi.x, logphi, w)
}
\arguments{
\item{x}{scale or vector; perp. distance observations}

\item{hr}{hazard rate function}

\item{ystart}{max forward distance at which could possibly detect animal (see details).}

\item{pi.x}{perpendicular distance density distribution}

\item{w}{perpendicular truncation distance.}

\item{y}{scalar or vector; forward distance observations}

\item{pars}{c(b,logphi); hazard rate and density log-parameters in a vector (see details).}

\item{length.b}{length of the hazard rate parameter vector}
}
\value{
negative log likelihood for forward distance, \code{y} and perpendicular distance \code{x}.
}
\description{
Returns product of perp dist det prob px() and animal dbn pi.x
}
\details{
Must to ensure the hazard function has decayed to (very close to) zero by \code{ystart}.
The parameter vector, \code{pars}, must be passed in with two parameters for the hazard rate first,
then two parameters for perpendicular density gradient \eqn{\pi(x)} i.e. \code{c(b,logphi)}.
}
\examples{
p.pi.x(x,b,hr,ystart,pi.x,logphi,w)
}

