% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{phat}
\alias{phat}
\title{Calculate effective strip width}
\usage{
phat(fit = NULL, w = NULL, hr = NULL, b = NULL, ystart = NULL,
  pi.x = NULL, logphi = NULL)
}
\arguments{
\item{w}{truncation distance}

\item{hr}{function describing the hazard rate}

\item{b}{hazard rate parameter vector}

\item{ystart}{maximum forward distance}

\item{pi.x}{function describing the perpendicular distance density distribution}

\item{logphi}{parameters for pi.x (some maybe logged)}

\item{fit=NULL}{alternatively just pass in an object resulting from a call of \link{fityx}.}
}
\value{
Effective strip widith \eqn{\hat p}
}
\description{
Calculate effective strip width, \eqn{\hat p} for a given hazard rate function and 
perpendicular density distribution.
}
\examples{
phat(w=1,hr=h2,b=log(c(0.75,1)),ystart=4,pi.x=pi.norm,logphi=c(0.5,log(0.2)))
}

