% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{phatInterval}
\alias{phatInterval}
\title{Calculates coverage probabilities of \eqn{\hat p}}
\usage{
phatInterval(fit, type = "LOGNORM", interval = 0.95)
}
\arguments{
\item{fit}{object resulting from a call of \code{\link{fityx}} (see details)}

\item{type}{\code{LOGNORM} log-normal confidence intervals; \code{NORM} normal confidence intervals.}

\item{interval}{the interval used to determine coverage probability}
}
\value{
a data frame of p phat, and CV[phat], interval, lower bound returned.
}
\description{
Calculate coverage probabilities of \eqn{\hat p} using the delta method and assuming a
log-normal error distribution.
}
\details{
In the call of \code{\link{fityx}} in the \code{fit} argument must have \code{hessian=TRUE}
}
\examples{
\dontrun{
ystart=4;w=1
hr=h2; b=log(c(0.75,1))
pi.x=pi.norm; logphi=c(0.5,log(0.2))
N=50 #true number of animals
#generate some observations
simDat=simXY(N=N,pi.x=pi.x,logphi=logphi,
hr=hr,b=b,w=w,ystart=ystart)
x=simDat$locs$x; y=simDat$locs$y
est.yx=fityx(fityx(y,x,b,hr,ystart,pi.x,logphi,w)
phatInterval(fit=est.yx,interval=0.95)
}
}
\seealso{
\code{\link{phat}} \code{\link{fityx}}
}

