% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{pi.chnorm}
\alias{pi.chnorm}
\title{Complementary half-normal form for perpendicular animal density function}
\usage{
pi.chnorm(x, logphi, w)
}
\arguments{
\item{x}{prependicular trackline distance}

\item{logphi}{numeric vector; parameters, some of which may be logged}

\item{w}{perpendicular truncation distance}
}
\value{
\eqn{\pi(x)} animal density at distance x
}
\description{
Complementary half-normal distribution of perpendicular animal density.
Truncation occurs at x=w (the perpendicular truncation distance)
}
\examples{
plot(seq(0,1,length=100),pi.hnorm(x=seq(0,1,length=100),logphi=0.5,w=1),
type='l',xlab='Perp. distance, x',ylab=expression(pi(x)))
}

