% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{pi.hr1}
\alias{pi.hr1}
\title{Perpendicular animal density function calulated from hazard rate \code{h1}}
\usage{
pi.hr1(x, logphi, w)
}
\arguments{
\item{x}{perpendicular distance.}

\item{logphi}{vector of parameters of \code{\link{h1}} (on log scale).}

\item{w}{perpendicular truncation distance}
}
\value{
Animal density at x calculated from hazard rate \code{\link{h1}}
}
\description{
Calculates perpendicular animal density \eqn{\pi(x)} using the prependicular 
distance function of the hazard rate function \eqn{k(x,y)=a*r^{-b}} on page 36  of Hayes and Buckland (1983) 
from the hazard function given on page 37, of that paper (and implemented 
in function \code{\link{hr1.to.p}}).
}
\examples{
logphi=log(c(0.01,1.01))
x=seq(0,1,length=50)
plot(x,pi.hr1(x,logphi,w=1),type="l",xlab="Perpendicular distance",
ylab=expression(pi(x)))
}
\seealso{
\code{\link{h1}} \code{\link{hr1.to.p}}
}

