% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{pi.hr2}
\alias{pi.hr2}
\title{Perpendicular animal density function calulated from hazard rate \code{hr2}}
\usage{
pi.hr2(x, logphi, w)
}
\arguments{
\item{x}{perpendicular distance.}

\item{logphi}{vector of parameters of \code{\link{h1}} (on log scale).}

\item{w}{perpendicular truncation distance}
}
\value{
Animal density at x calculated from hazard rate \code{\link{h1}}
}
\description{
Calculates perpendicular animal density \eqn{\pi(x)} using the prependicular
distance function of the hazard rate function \eqn{k(r,y)=a \sqrt{(r^2-y^2)}/r^{(b+1)}; b>2}
on page 36  of Hayes and Buckland (1983) #' from the hazard function given on page 37,
of that paper (and implemented in function \code{\link{hr1.to.p}}).
}
\examples{
x=seq(0,1,length=50)
plot(x,pi.hr2(x,logphi=c(-0.2876821, -2.3025851),w=1),
xlab='x',ylab=expression(pi(x)),type='l')
}
\seealso{
\code{\link{h1}} \code{\link{h2}} \code{\link{pi.hr1}}
}

