% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{pi.norm}
\alias{pi.norm}
\title{Truncated normal form for perpendicular animal density function}
\usage{
pi.norm(x, logphi, w)
}
\arguments{
\item{x}{prependicular trackline distance}

\item{logphi}{numeric vector; parameters, some of which may be logged}

\item{w}{perpendicular truncation distance}
}
\value{
\eqn{\pi(x)} animal density at distance x
}
\description{
Truncated normal distribution of perpendicular animal density.
Truncation occurs at x=0 (on the track line) and x=w (perpendicular truncation distance)
}
\examples{
plot(seq(0,1,length=100),pi.norm(x=seq(0,1,length=100),logphi=c(0.5,log(0.3)),w=1))
}

