% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{plotFit}
\alias{plotFit}
\title{Plot the 2D fit of a model}
\usage{
plotFit(fit, ...)
}
\arguments{
\item{fit}{object resulting from a call of \link{fityx}}

\item{...}{other parameters passed into \link{plotSim}}
}
\description{
Plot the 2D fit of a model resulting from a call of \link{fityx}
}
\details{
This function is a wrapper for \link{plotSim}.
}
\seealso{
\link{plotSim} \link{fityx}
}

