% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{plotfit.y}
\alias{plotfit.y}
\title{Plot fitted f(y) and forward distance distribution}
\usage{
plotfit.y(y = NULL, x = NULL, est, nclass = 10, breaks = NULL,
  plot = TRUE, dotitle = FALSE, lineonly = FALSE, nint = 100,
  max.obs = TRUE, add = FALSE, ...)
}
\arguments{
\item{y}{forward distance observations (if NULL, uses est$dat$y)}

\item{x}{perpendicular distance observations (if NULL, uses est$dat$x)}

\item{est}{return from a call of \code{\link{fityx}}}

\item{nclass}{number of histogram classes}

\item{breaks}{break points passed to hist (overrides nclass if not NULL)}

\item{plot}{boolean, plot results}

\item{lineonly}{if TRUE plots only f(y), else plots histogram of forward distances too}

\item{nint}{number of intervals to use in calculating f(y)}

\item{max.obs}{If TRUE, plots only up to maximum observed forward distance, else plots
up to est$ystart (the forward distance beyond which detection is assumed impossible).}

\item{add}{if TRUE, adds line to existing plot, else creates new plot. Only applicable
if lineonly==TRUE.}

\item{...}{other parameters passed to \code{hist} and \code{plot} (need to separate these two!)}
}
\value{
Invisibly returns a list with these elements
\code{$gridx} = x values used in plotting
\code{$fy.x} = Unscaled f(y|x) for all the xs observed
\code{$fy.} = Unscaled mean of f(y|x) for all the xs observed
\code{$scaled. fy.} = Mean of f(y|x), scaled to integrate to 1
}
\description{
Plot f(y) and forward distance distribution
resulting from a call of \code{\link{fityx}}.
}
\details{
Plot f(y) and forward distance distribution resulting from a call of
\code{\link{fityx}}, optionall with overlaid histogram of foward distances.
Invisibly returns various f(y)'s, as detailed below.
}
\examples{
\dontrun{
ystart=4;w=1
hr=h2; b=log(c(0.75,1))
pi.x=pi.norm; logphi=c(0.5,log(0.2))
N=50 #true number of animals
#generate some observations
simDat=simXY(N=N,pi.x=pi.x,logphi=logphi,
hr=hr,b=b,w=w,ystart=ystart)
x=simDat$locs$x; y=simDat$locs$y
est.yx=fityx(y,x,b,hr,ystart,pi.x,logphi,w)
plotfit.y(y,x,est.yx,nclass=10)
}
}
\seealso{
\code{\link{fityx}}
}

