% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{plotSim}
\alias{plotSim}
\title{Plot the simulated positions}
\usage{
plotSim(simDat, nclass = 10, xlab = "", ylab = "", image = FALSE, ...)
}
\arguments{
\item{simDat}{object from a call of \code{\link{simXY}}}

\item{nclass}{number of bins in x and y histograms}

\item{xlab}{x-axis label for sightings scatter plot}

\item{ylab}{y-axis label for sightings scatter plot}

\item{image}{Boolean \code{TRUE} image background on sightings scatter plot.}

\item{...}{other arguments to be passed into \code{\link{plotfit.y}}}
}
\description{
Plots the simulated data.
}
\examples{
n=100;ymin=0.01;ymax=5;W=1
b=log(c(0.75,1));logphi=c(0.5,log(0.2))
simDat=sim.n(n,ymin,ymax,W,h2,b,pi.norm,logphi)
plotSim(simDat=simDat,nclass=12)
}
\seealso{
\code{\link{simXY}} \code{\link{plotSim}} \code{\link{plotfit.y}}
}

