% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{plotSimNotUsed}
\alias{plotSimNotUsed}
\title{Plot the simulated positions}
\usage{
plotSimNotUsed(simDat)
}
\arguments{
\item{simDat}{object from a call of \code{\link{simXY}}}
}
\description{
Plots the simulated data.
}
\examples{
ystart=4;w=1
hr=h2; b=log(c(0.75,1))
pi.x=pi.norm; logphi=c(0.5,log(0.2))
N=500 #true number of animals
#generate some observations
simDat=simXY(N=N,pi.x=pi.x,logphi=logphi,
hr=hr,b=b,w=w,ystart=ystart)
plotSimNotUsed(simDat)
}

