% Generated by roxygen2 (4.0.2): do not edit by hand
\name{poisint}
\alias{poisint}
\title{NHPP intensity calculation in region}
\usage{
poisint(y, x, ymin, ymax, hfun, b, pi.x, logphi, W, lscale = 1)
}
\arguments{
\item{y}{forward distances at which to calculate intensities (vector)}

\item{x}{perpendicular distances at which to calculate intensities (must
be same length as y).}

\item{ymin}{smallest forward distance}

\item{ymax}{largest forward distance}

\item{hfun}{detection hazard function}

\item{b}{vector of detection hazard function parameters}

\item{pi.x}{perpendicular distance distribution function}

\item{logphi}{vector with log of pi.x parameters}

\item{W}{perpendicular truncation distance}

\item{lscale}{output of \code{\link{calc.lpars}} (object of class ``ppscale'')}
}
\value{
NHPP intensities at all \code{(x,y)}s input.
}
\description{
Calculates NHPP intensity in given region, using
             given detection hazard and perp. dist. distribution.
}
\details{
Calculates survival model pdf of forward distance \code{y}, given
perpendicular distance \code{x} (\code{f(y|x).}), multiplies this by the
perpendicular distance pdf \code{pi.x} and then scales it by multiplying
by lscale (in order to get some total intensity - typically that to
generate some expected sample size). See \code{\link{calc.lpars}} for
details of the scaling.

This function is called by \code{rpoispp} inside \code{sim.n} to generate
samples from NHPPs.
}
\examples{
\dontrun{
n=100;ymin=0.01;ymax=5;W=2
b=log(c(0.75,1));logphi=c(0.5,log(0.3))
nf=100
ys=seq(ymin,ymax,length=nf)
intscale=calc.lpars(n,ymin,ymax,W,h2,b,pi.norm,logphi)
f=poisint(ys,rep(0,nf),ymin=ymin,ymax=ymax,hfun=h2,b=b,pi.x=pi.norm,logphi=logphi,W=W,lscale=intscale$lscale)
plot(ys,f,type="l",xlab="Forward distance (y)",ylab="f(y)")
}
}

