% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{primate.dat}
\alias{primate.dat}
\title{Primate line transect survey data.}
\format{A list with elements x (perpendicular distance) and y (forward distance).}
\source{
We are gerateful to Matthew Nowak from the Sumatran Orangutan Conservation
Programme (SOCP) for allowing us to use the primate survey data from the Jantho
Reintroduction Station. The initial survey was developed by Matthew Nowak and Serge
Wich (Liverpool John Moores University) and then undertaken by the SOCP with funding
from Chester Zoo. .
}
\usage{
primate.dat
}
\description{
Locations relative to the observer of 127 detections of primates from a
visual survey conducted by three sets of trained observers walking previously cut line
transects in primary tropical rainforest.
}
\examples{
data(primate.dat)
}

