% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{simnhPP}
\alias{simnhPP}
\title{Simulate forward distances}
\usage{
simnhPP(x, b, ystart, hr, miss = TRUE, ylo = 1e-05)
}
\arguments{
\item{x}{scale or vector; perpendicular distance observations}

\item{b}{two-element vector of hazard rate parameters, some of whihc may be logged}

\item{ystart}{max forward distance at which could possibly detect animal (see details).}

\item{hr}{hazard rate function}

\item{miss}{If TRUE, allows animals not detected by y=0 to have no y.}

\item{ylo}{minimum forward distance.}
}
\value{
vector of y distances if \code{miss=TRUE} missed animals represented by -999, otherwise
no animals are missed.
}
\description{
Simulate non-homogeneous Poisson Process data by solving inverse CDF.
This function has been replaced by \code{\link{sim.n}}.
}
\details{
\code{miss} allows a fixed number of animal to be simulated i.e. known $N$ in strip,
or $n$ where animals may have been missed.
}
\examples{
ystart=4
hr=h2; b=log(c(0.75,1))
X=runif(50,0,1)
Y=simnhPP(x=X,b=b,ystart=ystart,
hr=h2,miss=TRUE)
length(Y[Y!=-999])
Y=simnhPP(x=X,b=b,ystart=ystart,
hr=h2,miss=FALSE)
length(Y[Y!=-999])
}
\seealso{
\code{\link{simXY}}
}

