% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2DLTsimulator.R
\name{simulator}
\alias{simulator}
\title{Simulate observations and fit a model}
\usage{
simulator(N = NULL, fit, n = NULL, ymin = NULL, optimx = FALSE,
  debug = FALSE)
}
\arguments{
\item{optimx}{use the \code{\link{optimx::optimx()}} function for fitting, Default is FALSE.}

\item{n=NULL}{if n is a positive integer, simulator will run until number of observations equals n}

\item{ymin=NULL}{if n is a positive real number, simulator will discard simulated positions <ymin}

\item{debug=FALSE}{additional output for debugging. See details}

\item{...}{other arguments to be passed into \link{\code{simXY}}}
}
\description{
Simulate observations and fit a model
}
\details{
If n=NULL and ymin=NULL then use \link{\code{simXY}}.  When n!=NULL this function iteratively calls \link{\code{simXY}} until n simulated observations are reached.
When ymin!=NULL this function discards simulated data with Y-coordinates <ymin.  It is intended that \code{ymin}
is calculated by a call of \code{yDistForg} to determine the y distance at which g(y)= a required detection probability
If \code{debug=TRUE} a list object including the model fit is returned.
}
\note{
Simulate observations from a hazard rate and perpendicular distribution and fit a model to the simulated observations.
Currently the function uses the same hazard rate, prependicular density distribuion and parameters for the simulated data and initial values for fitting.
}
\examples{
system.time({out=simulator(n=60,fit=fit.n.ip1,optimx=FALSE)})
}
\author{
Martin J. Cox
}

