% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2DLTfunctions.r
\name{simXY}
\alias{simXY}
\title{Simulate sightings given a known perpendicular density distribution and hazard function}
\usage{
simXY(N, pi.x, logphi, hr, b, w, ystart, xSampL = 1000,
  discardNotSeen = TRUE, ...)
}
\arguments{
\item{N}{animal population}

\item{pi.x}{function describing the perpendicular distance density distribution}

\item{logphi}{parameters for pi.x (some maybe logged)}

\item{hr}{function describing the hazard rate}

\item{b}{hazard rate parameter vector}

\item{w}{truncation distance}

\item{ystart}{max forward distance at which could possibly detect animal (see details).}

\item{xSampL}{length of x-dimension vector to sample perpendicular distances from.}

\item{discardNotSeen}{boolean; discard individuals not detected.  See details}

\item{...}{arguments to be passed into \code{\link{simnhPP}}}
}
\value{
list of \code{$locs} x and y coordinates for simulated sightings and
\code{$settings} simulation settings.
}
\description{
Simulates sightings from a known population given a density distribution
and hazard rate.  This function has been replaced by \code{\link{sim.n}}.
}
\details{
if \code{discardNotSeen=FALSE} individuals that are not detected are
assigned y-dimension distances = -999, otherwise \code{discardNotSeen=TRUE}
invididuals are removed and not returned
}
\examples{
ystart=4;w=1
hr=h2; b=log(c(0.75,1))
pi.x=pi.norm; logphi=c(0.5,log(0.2))
N=50 #true number of animals
#generate some observations
simDat=simXY(N=N,pi.x=pi.x,logphi=logphi,
hr=hr,b=b,w=w,ystart=ystart)
x=simDat$locs$x; y=simDat$locs$y
}

