% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTsimulator.R
\name{simXY2}
\alias{simXY2}
\title{Data simulator}
\usage{
simXY2(pi.x, logphi, hr, b, w, ystart, xSampL = 1000, discardNotSeen = TRUE,
  n = NULL, ymin = NULL, ...)
}
\arguments{
\item{...}{other arguments to be passed into \link{\code{simXY}}}

\item{n=NULL}{if n is a positive integer, simulator will run until number of observations equals n}

\item{ymin=NULL}{if n is a positive real number, simulator will discard simulated positions <ymin}
}
\description{
Simulate a fixed number of observations potentially when \code{g(0)<1}.  This is essentially a wrapper function for \code{\link{simXY}}
}
\details{
If n=NULL and ymin=NULL then use \link{\code{simXY}}.  When n!=NULL this function iteratively calls \link{\code{simXY}} until n simulated observations are reached.
When ymin!=NULL this function discards simulated data with Y-coordinates <ymin.  It is intended that \code{ymin}
is calculated by a call of \code{yDistForg} to determine the y distance at which g(y)= a required detection probability
}
\examples{
\dontrun{
minY=yDistForg(gy=0.8,ymax=dfit.hn$ystart,b=dfit.hn$b,hfun=dfit.hn$hr)
minY
simDatg0dot8=NULL
simDatg0dot8=simXY2(pi.x=dfit.hn$pi.x,logphi=dfit.hn$logphi,
                   hr=dfit.hn$hr,b=dfit.hn$b,w=dfit.hn$w,ystart=dfit.hn$ystart,
                   n=2000,ymin=minY)
min(simDatg0dot8$y)
}
}

