% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Sy}
\alias{Sy}
\title{Calculates survivor function}
\usage{
Sy(x, y, ymax, b, hfun)
}
\arguments{
\item{x}{perpendicular distance (scalar)}

\item{y}{forward distance  (scalar)}

\item{ymax}{largest forward distance}

\item{b}{vector of detection hazard function parameters}

\item{hfun}{detection hazard function}
}
\value{
Probability of making it from forward distance
\code{ymax} to \code{y} without being detected.
}
\description{
Calculates survivor function to forward distance \code{y},
for given perpendicular distance \code{x} and given forward distance range.
}
\details{
Calculates probability of making it from forward distance
\code{ymax} to \code{y} without being detected.
}
\examples{
\dontrun{
ymax=5
Sy(0,0.1,ymax,b,h2)
}
}

