% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2DLTsimulator.R
\name{yDistForg}
\alias{yDistForg}
\title{Determine minimum forward distance for simulated data that will yield a given \code{(g0)<1}}
\usage{
yDistForg(gy, ymax, b, hfun, tolerance = 1e-04, ...)
}
\arguments{
\item{gy}{desired detectability at \code{y=0}}

\item{ymax}{largest forward distance}

\item{b}{vector of detection hazard function parameters}

\item{hfun}{detection hazard function}

\item{tolerance}{passed into \code{optimize()}}
}
\description{
This is a helper function for simulating data with \code{g(0)<1}.  The function
 calculates a minimum forward distance, \code{y_{min}}, to truncate simulated data, i.e. simulated sightings
 with \code{y_i<y_{min}} are deleted.
}
\details{
The idea behind this function is that simulated data with \code{y_i<y_{min}} would got have been
 detected at a given \code{g(0)<1}.  The \code{y_{min}} for a given \code{g(0)<1} is solved using
 \code{optimise()}.  The resulting \code{y_{min}} is substracted from the remaining simulated observations.
}
\examples{
\dontrun{
###No bump.
minY=yDistForg(gy=0.8,ymax=dfit.hn$ystart,b=dfit.hn$b,hfun=dfit.hn$hr)
minY
1-Sy(y=minY,x=0,ymax=dfit.hn$ystart,b=dfit.hn$b,hfun=dfit.hn$hr)
}
}

